Fix Eigen SEGV caused by missing return values
In C++, declaring a return value but not returning anything is
unspecified behavior. Several functions/methods in this library are
doing this, and in my case, it's causing a nasty SEGV to occur in
unrelated Eigen code (probably because the program counter is falling of
the end of the function and executing the unrelated Eigen code that
happens to be below!). This is bad, so let's actually return (unused)
values from these functions.

Note: using gcc version 11.4.0