badger: remove dependency on livox_ros_driver
We aren't using a Livox lidar, so why bother putting the driver and SDK
into the yocto image? Instead, let's make the LIO library not require
the Livox-specific driver by removing all the code that uses it. We
don't need these functions because we are using the versions of them
that support PointCloud2 data.

This is a Badger-specific change, so I won't include it in the upstream
PR.