badger: running w/ Hesai QT128 (working WIP)
My objective was to evaluate this open-source LIO-Livox lidar odometry
algorithm to see how well it works. This required making it run with our
Hesai top-front lidar. Here is a sloppy commit with those changes, some
half-baked evaluation infrastructure, and a braindump README to make
this useful for posterity.

This is a Badger-specific change, so I don't plan to include it in the
upstream PR.